:sparkles: added prismic service with GetSettings method
As I wrote in the comment, tests for this method don't pass. Unfortunately, prismic classes can't be mocked. To fix that I would need to create wrappers which I would be able to use in tests, however this task will enlarge without any control. I will get back to this in my spare time.