:sparkles: added get Dilemma implementation
It is really harmful that I wasn't able to add tests to this method, as there are a lot of to test here. But let's at least describe.

Prismic structure for:
- Dilemma document is the root one, contains some description + link to first step
- DilemmaStep document is a child document, also contains some details + list of user choices, which in fact are links to other DilemmaSteps

So in other words, Prismic structure reflects regular flowcharts - it is a tree, but a tricky one.
We need to bear in mind that flowchart can point to the same destination steps from different origin roots, but what is more important it can create loops.
Simple and naive implementation for cascade steps resolving is not an option here.

Above code creates one root model with the list of all unique steps that were used in the tree.
FE code will be able to match choices with correct steps by step id.